pyproject.toml
[tool.poetry]
name = "test-bot"
version = "0.1.0"
description = "Bot for portfolio"
authors = ["Alexander"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.27"
aiogram = "^3.3.0"
asyncpg = "^0.29.0"
python-dateutil = "^2.8.2"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
